Moved data sources table to new component
